Move number displayer closer to trust, where ot copmes from
