+++ src/app/shared/keml/components/helper/trust/trust.component.spec.ts
@@ -1,7 +1,7 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { TrustComponent } from './trust.component';
-import {NumberDisplayer} from "@app/core/utils/number-displayer";
+import {NumberDisplayer} from "@app/shared/keml/utils/number-displayer";
 
 describe('TrustComponent', () => {
   let component: TrustComponent;
